[JENKINS-40105_Enable_all_folder.js_test_cases] remove async
